Add a 404 page
